Rename test_config_encryption.py to test_encryption.py
This change is done because the code being tested is in encryption.py, not in config_encryption.py
