+++ monkey/tests/unit_tests/monkey_island/cc/services/utils/test_encryption.py
